+++ src/main/groovy/ru/vyarus/gradle/plugin/python/PythonPlugin.groovy
@@ -4,6 +4,7 @@ import groovy.transform.CompileStatic
 import groovy.transform.TypeCheckingMode
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.tasks.Delete
 import org.gradle.api.tasks.TaskProvider
 import ru.vyarus.gradle.plugin.python.task.BasePythonTask
 import ru.vyarus.gradle.plugin.python.task.CheckPythonTask
@@ -26,6 +27,7 @@ import ru.vyarus.gradle.plugin.python.task.pip.PipUpdatesTask
  * @since 11.11.2017
  */
 @CompileStatic
+@SuppressWarnings('DuplicateStringLiteral')
 class PythonPlugin implements Plugin<Project> {
 
     @Override
@@ -62,6 +64,16 @@ class PythonPlugin implements Plugin<Project> {
         project.tasks.register('pipList', PipListTask) {
             it.with {
                 description = 'Show all installed modules'
+            }
+        }
+
+        project.tasks.register('cleanPython', Delete) {
+            it.with {
+                group = 'python'
+                description = 'Removes existing python environment (virtualenv)'
+                delete extension.envPath
+                // ignore if virtualenv not created
+                onlyIf { extension.scope.ordinal() >= PythonExtension.Scope.VIRTUALENV_OR_USER.ordinal() }
             }
         }